release: releae v1.0.0-gama
Release published v1.0.0-gama
